= Refactor social api directory structure
